Delete action gets confirm dialog
The rudimentary one.

+++ v3/app/controllers/genericEditController.js
@@ -58,6 +58,11 @@ define([ '../ngmodule' ], function(appModule) {
         };
 
         $scope.deleteItem = function() {
+
+            if(!confirm('Delete the current item?')){
+                return;
+            }
+
             var item = $scope.editing.original;
             item.destroy({
                 success : function() {
